# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.BankStatsResponse:
    properties:
      bank_id:
        example: x9y8z7w6v5u4t3s2
        type: string
      mastery:
        example: 42
        type: integer
      question_stats:
        items:
          $ref: '#/definitions/api.QuestionStatsResponse'
        type: array
      total_questions:
        example: 10
        type: integer
    type: object
...
  api.CreateBankRequest:
    properties:
      bank_type:
        example: theory
        type: string
      category_id:
        example: a1b2c3d4e5f6g7h8
        type: string
      difficulty:
        example: medium
        type: string
      language:
        example: go
        type: string
      subject:
        example: Go concurrency patterns
        type: string
    type: object
  api.CreateBankResponse:
    properties:
      bank_type:
        example: theory
        type: string
      category_id:
        example: a1b2c3d4e5f6g7h8
        type: string
      difficulty:
        example: medium
        type: string
      id:
        example: x9y8z7w6v5u4t3s2
        type: string
      language:
        example: go
        type: string
      mastery:
        example: 0
        type: integer
      question_count:
        example: 5
        type: integer
      subject:
        example: Go concurrency patterns
        type: string
    type: object
...
  api.GetBankResponse:
    properties:
      bank_type:
        example: theory
        type: string
      category_id:
        example: a1b2c3d4e5f6g7h8
        type: string
      difficulty:
        example: medium
        type: string
      id:
        example: x9y8z7w6v5u4t3s2
        type: string
      language:
        example: go
        type: string
      mastery:
        example: 42
        type: integer
      questions:
        items:
          $ref: '#/definitions/api.QuestionResponse'
        type: array
      subject:
        example: Go concurrency patterns
        type: string
    type: object
...
  api.QuestionResponse:
    properties:
      expected_answer:
        example: A goroutine is a lightweight thread managed by the Go runtime.
        type: string
      grading_prompt:
        example: Be strict about mentioning the Go scheduler.
        type: string
      id:
        example: q1w2e3r4t5y6u7i8
        type: string
      mastery:
        example: 75
        type: integer
      subject:
        example: What is a goroutine?
        type: string
      times_answered:
        example: 3
        type: integer
      times_correct:
        example: 2
        type: integer
    type: object
  api.QuestionStatsResponse:
    properties:
      mastery:
        example: 75
        type: integer
      question_id:
        example: q1w2e3r4t5y6u7i8
        type: string
      times_answered:
        example: 3
        type: integer
      times_correct:
        example: 2
        type: integer
    type: object
...
  api.UpdateBankCategoryRequest:
    properties:
      category_id:
        example: a1b2c3d4e5f6g7h8
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal learning companion — create question banks, practice, and
    let AI grade your answers.
  title: Remaimber-it API
  version: "1.0"
paths:
  /banks:
    get:
      description: Returns all question banks across all categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.CreateBankResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all banks
      tags:
      - Banks
    post:
      consumes:
      - application/json
      description: Create a new question bank within a category.
      parameters:
      - description: Bank to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateBankRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CreateBankResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a question bank
      tags:
      - Banks
  /banks/{bankID}:
    delete:
      description: Delete a bank and cascade-delete all its questions and stats.
      parameters:
      - description: Bank ID
        in: path
        name: bankID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a question bank
      tags:
      - Banks
    get:
      description: Returns a question bank with all its questions and their stats.
      parameters:
      - description: Bank ID
        in: path
        name: bankID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetBankResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a question bank
      tags:
      - Banks
  /banks/{bankID}/category:
    patch:
      consumes:
      - application/json
      description: Move a question bank to a different category (or set to null).
      parameters:
      - description: Bank ID
        in: path
        name: bankID
        required: true
        type: string
      - description: New category
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateBankCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateBankResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update bank category
      tags:
      - Banks
...
  /banks/{bankID}/stats:
    get:
      description: Returns mastery and per-question statistics for a bank.
      parameters:
      - description: Bank ID
        in: path
        name: bankID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BankStatsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get bank stats
      tags:
      - Banks
...
swagger: "2.0"